Updated position field type for Section model of privacy_policy app

+++ privacy_policy/models.py
@@ -11,7 +11,7 @@ def section_default_position() -> int:
 class Section(models.Model): # type: ignore [django-manager-missing]
 	title = models.CharField(_('Заголовок'), max_length=255)
 	text = models.TextField(_('Текст'))
-	position = models.IntegerField(_('Позиция'), blank=True, default=section_default_position)
+	position = models.PositiveSmallIntegerField(_('Позиция'), blank=True, default=section_default_position)
 
 	class Meta(TypedModelMeta):
 		ordering = ('position',)
